blogapi/blog_proj/posts/serializers.py
from django.contrib.auth import get_user_model
from django.contrib.auth.models import User

from rest_framework import serializers

from .models import Post

class PostSerializer(serializers.ModelSerializer):

    class Meta:
        model = Post
        fields = ('id', 'title', 'body', 'created_at', 'author',)

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = get_user_model()
        fields = ("id", "username",)