blogapi/blog_proj/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.1
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
types-cryptography==3.3.23
urllib3==1.26.12
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.1
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
types-cryptography==3.3.23
urllib3==1.26.12
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-rest-auth==0.9.5
django-rest-auth-forked==0.9.5
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.1
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
types-cryptography==3.3.23
urllib3==1.26.12
